+++ src/main/java/frc/robot/subsystems/vision/Vision.java
@@ -1,14 +1,19 @@
 /* (C) Robolancers 2025 */
 package frc.robot.subsystems.vision;
 
-import edu.wpi.first.epilogue.Logged;
-import edu.wpi.first.wpilibj2.command.SubsystemBase;
+import frc.robot.subsystems.vision.VisionConstants.CameraConfig;
+import frc.robot.util.VirtualSubsystem;
 
-// for specific code structure examples, refer to
-// https://github.com/BananasAmIRite/offseason-2025/tree/code-structure/src/main/java/frc/robot/subsystems
+public class Vision extends VirtualSubsystem {
 
-// `@Logged` must be placed in front of every class you wish to log
-// for more info about logging, visit:
-// https://docs.wpilib.org/en/stable/docs/software/telemetry/robot-telemetry-with-annotations.html#new-for-2025-robot-telemetry-with-annotations-java-only
-@Logged
-public class Vision extends SubsystemBase {}
+  private final PoseEstimatorCamera estimatorCamera;
+
+  public Vision(CameraConfig estimatorConfig) {
+    this.estimatorCamera = new PoseEstimatorCamera(estimatorConfig);
+  }
+
+  @Override
+  public void periodic() {
+    estimatorCamera.update();
+  }
+}
