+++ src/main/java/frc/robot/subsystems/vision/VisionConstants.java
@@ -1,5 +1,5 @@
 /* (C) Robolancers 2025 */
-package frc.robot.vision;
+package frc.robot.subsystems.vision;
 
 import edu.wpi.first.apriltag.AprilTagFieldLayout;
 import edu.wpi.first.apriltag.AprilTagFields;
@@ -8,7 +8,7 @@ import edu.wpi.first.math.geometry.Transform3d;
 public class VisionConstants {
   public static record CameraConfig(String cameraName, Transform3d robotToCamera) {}
 
-  public static final CameraConfig kCameraConfig427 = new CameraConfig("camera", new Transform3d());
+  public static final CameraConfig k427CameraConfig = new CameraConfig("camera", new Transform3d());
 
   public static final AprilTagFieldLayout kAprilTagFieldLayout =
       AprilTagFieldLayout.loadField(AprilTagFields.k2025Reefscape);
